requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.13.2
